% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportCreation.R
\name{Forecast_report}
\alias{Forecast_report}
\alias{Forecast_report.data.frame}
\title{Reporting the information of Forecasting}
\usage{
Forecast_report(...)

\method{Forecast_report}{data.frame}(
  RESULTS,
  OverallTimePoint,
  DescriptionDF = NULL,
  Total = NULL,
  HistoricalCurve = NULL,
  output_format = "html",
  output_file = NULL,
  output_dir = NULL,
  browse = TRUE,
  ...
)
}
\arguments{
\item{...}{arguments to be passed to methods.}

\item{output_format}{report output type. Choose either "pdf" and "html".
"pdf" create pdf file by knitr::knit().
"html" create html file by rmarkdown::render().}

\item{output_file}{name of generated file. default is NULL.}

\item{output_dir}{name of directory to generate report file. default is tempdir().}

\item{.data}{a tbl_dbi.}

\item{DataDict}{dataframe specifying variable, variable description, and variable type.}

\item{DateName}{Name of date variable in dataset}

\item{datetofilteron}{a Date. The date to separate training and test for PSI calculation.}

\item{saveoutputs}{a character path string. If populated, will save png and csv outputs to this folder.}
}
\description{
The Forecast_report() report the information of Exploratory
data analysis for object inheriting from data.frame.
}
\details{
Generate generalized data quality reports automatically.
You can choose to output to pdf and html files.
This is useful for a data frame with a large number of variables
than data with a small number of variables.
}
\section{Reported information}{

The Data Quality process will report the following information:

\itemize{
  \item Missing and Unique Values
  \itemize{
    \item Missing Values
    \item Unique Values
    \item Percent Zero Values
    \item Percent Negative Values
    \item Percent Outliers
  }
  \item Descriptive Statistics
  \item Categorical Breakdown
  \itemize{
    \item Percent of Data in each level
  }
  \item Histograms
  \item Population Stability Index
  \item Trended Numerical Variables
  \item Trended Categorical Variables
  \item Correlation Matrix
}

See vignette("Forecast_repoort") for an introduction to these concepts.
}

\examples{
\donttest{

carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Forecast_report("US", output_format = "html")

# create html file. file name is EDA.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Forecast_report(output_format = "html", output_file = "EDA.html")


# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Forecast_report("Sales", output_format = "html")

# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Forecast_report(output_format = "html")

# create html file. file name is EDA2.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Forecast_report(output_format = "html", output_file = "EDA2.html")
}

}
\seealso{
\code{\link{Forecast_report.data.frame}}.
}
