% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluation.R
\docType{package}
\name{ModelEvaluation-package}
\alias{ModelEvaluation}
\alias{ModelEvaluation-package}
\title{ModelEvaluation: Tools for Data Diagnosis, Exploration, Transformation}
\description{
ModelEvaluation provides data diagnosis, data exploration and transformation of variables
during data analysis.
}
\details{
It has two main goals:

\itemize{
\item When data is acquired, it is possible to judge whether data is erroneous
or to select a variable to be corrected or removed through data diagnosis.
\item Understand the distribution of data in the EDA process. We can also
understand the relationship between target variables and predictor variables
for the prediction model.
}

To learn more about ModelEvaluation, start with the vignettes:
`browseVignettes(package = "ModelEvaluation")`
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/nkreimold/ModelEvaluation/issues}
}

}
\author{
\strong{Maintainer}: Reimold Nicholas \email{nkreimold@gmail.com}

}
