% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{Qual_report.tbl_dbi}
\alias{Qual_report.tbl_dbi}
\title{Reporting the information of Data Quality for table of the DBMS}
\usage{
\method{Qual_report}{tbl_dbi}(
  .data,
  DataDict = NULL,
  DateName = NULL,
  datetofilteron = NULL,
  saveoutputs = NULL,
  output_format = "html",
  output_file = NULL,
  output_dir = tempdir(),
  in_database = FALSE,
  collect_size = Inf,
  ...
)
}
\arguments{
\item{.data}{a tbl_dbi.}

\item{DataDict}{dataframe specifying variable, variable description, and variable type.}

\item{DateName}{Name of date variable in dataset}

\item{datetofilteron}{a Date. The date to separate training and test for PSI calculation.}

\item{saveoutputs}{a character path string. If populated, will save png and csv outputs to this folder.}

\item{output_format}{report output type. Choose either "pdf" and "html".
"pdf" create pdf file by knitr::knit().
"html" create html file by rmarkdown::render().}

\item{output_file}{name of generated file. default is NULL.}

\item{output_dir}{name of directory to generate report file. default is tempdir().}

\item{...}{arguments to be passed to methods.}
}
\description{
The Qual_report() report the information of Exploratory
data analysis for object inheriting from the DBMS table through tbl_dbi
}
\details{
Generate generalized data quality reports automatically.
You can choose to output to pdf and html files.
This is useful for a data frame with a large number of variables
than data with a small number of variables.
}
\section{Reported information}{

The Data Quality process will report the following information:

\itemize{
  \item Missing and Unique Values
  \itemize{
    \item Missing Values
    \item Unique Values
    \item Percent Zero Values
    \item Percent Negative Values
    \item Percent Outliers
  }
  \item Descriptive Statistics
  \item Categorical Breakdown
  \itemize{
    \item Percent of Data in each level
  }
  \item Histograms
  \item Population Stability Index
  \item Trended Numerical Variables
  \item Trended Categorical Variables
  \item Correlation Matrix
}

See vignette("Qual_repoort") for an introduction to these concepts.
}

\examples{
\donttest{
library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Qual_report("US", output_format = "html")

# create html file. file name is EDA.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Qual_report(output_format = "html", output_file = "EDA.html")


# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Qual_report("Sales", output_format = "html")

# create html file. file name is EDA_Report.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Qual_report(output_format = "html")

# create html file. file name is EDA2.html
con_sqlite \%>\% 
  tbl("TB_CARSEATS") \%>\% 
  Qual_report(output_format = "html", output_file = "EDA2.html")
}

}
\seealso{
\code{\link{Qual_report.data.frame}}.
}
