% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitWebsite.R
\name{RenderWebsite}
\alias{RenderWebsite}
\alias{RenderWebsite.data.frame}
\title{Data Quality and Model Performance Report Generation}
\usage{
RenderWebsite(.data, ...)

\method{RenderWebsite}{data.frame}(
  Pathvars = NULL,
  basepath = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  ...
)
}
\arguments{
\item{...}{arguments to be passed to methods.}

\item{Pathvars}{a tbl_dbi. containing Folder names and paths of all data files(within basepath)}

\item{basepath}{a string for path where all metrics are located}

\item{output_dir}{name of directory to generate report file. default is tempdir().}
}
\description{
The RenderWebsite() produces reports for evaluating
performance of a two stage modeling process.
}
\details{
Knit together a bunch of model performance and data quality reports into a website. 
All report objects must have been previously generated
}
\section{Reported information}{

The Model Performance process will report the following information:

\itemize{
  \item Index- Summary Page
  \item Raw and Model Data Quality
  \item Performance for each segment
  \item Overall Performance
}

See vignette("Qual_repoort") for an introduction to these concepts.
}

\examples{
\donttest{
pathvars must contain names- used 'Data' if data quality and 'Performance' if model performance

paths can be duplicated, as long as everything has at least one common subdirectory 


Pathvars=data.frame(Names=c("Raw Data Active","Raw Data DQ",
"Secured Model Data Active","Secured Model Data DQ 30-60",
"Secured Model Data DQ 60-90","Secured Model Data DQ 90+",
"Unsecured Model Data Active","Unsecured Model Data DQ 30-60",
"Unsecured Model Data DQ 60-90","Unsecured Model Data DQ 90+",
"Secured Performance Active","Secured Performance DQ 30-60",
"Secured Performance DQ 60-90","Secured Performance DQ 90+",
"Unsecured Performance Active", "Unsecured Performance DQ 30-60",
"Unsecured Performance DQ 60-90","Unsecured Performance DQ 90+"),

paths=c( "RawData/Active/","RawData/DQ/",
"SECUREDActive0/","SECUREDDQ30/","SECUREDDQ60/","SECUREDDQ90/",
"UNSECUREDActive0/","UNSECUREDDQ30/","UNSECUREDDQ60/","UNSECUREDDQ90/" ,
"SECUREDActive0/","SECUREDDQ30/","SECUREDDQ60/","SECUREDDQ90/",
"UNSECUREDActive0/","UNSECUREDDQ30/","UNSECUREDDQ60/","UNSECUREDDQ90/" )
,stringsAsFactors = F)


basepath="S:/30000/Credit Policy/Nicholas/GitHub/CECL/CECL_CL_PD_EAD/ModelEval/ReportObjects2"

RenderWebsite(Pathvars,basepath,output_dir=basepath)


}
}
\seealso{
\code{\link{RenderWebsite.data.frame}}.
}
